+++ src/OutputBuilderClient/Interfaces/IBrokenImageTracker.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace OutputBuilderClient.Interfaces
+{
+    public interface IBrokenImageTracker
+    {
+        void LogBrokenImage(string path, Exception exception);
+
+        string[] AllBrokenImages();
+    }
+}
