+++ src/OutputBuilderClient/Services/BitlyUrlShortner.cs
@@ -6,6 +6,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using System.Web;
+using OutputBuilderClient.Interfaces;
 
 namespace OutputBuilderClient.Services
 {
